Use input figs (#123)
* Make figs includable in main document

Remove the capability to build them as standalone document but we can
include them in the main document.  It should simplify things down the
road.

* Replace tikx pics

* Remove figures compilation from makefile

No need to compile figs to pdf anymore, at least to build the books

* Delete svg figures from ebook build

* Create png for TikZ figures

    - Add export_figures back
    - Build a pdf from the the TikZ in standalone mode
    - Change the cleanup to deal with those changes
    - Remove trailing spaces..

* Replace centering by an environment

More LaTeX idomatic

* Increase clean_figures robustness

as we use temporary tex files (.tex.in) we can't run clean_figures twice in a
row

* Center TikZ figures

Because it looks better

* Remove png building

Seems to struggle on CI with \\\b or something, we don't absolutely need
it right now so let's get rid of it.

* Remove trgt_figures dependency for pdf and ebooks

This should not be needed now that we include the TikZ directly, only
needed for png generation.

* Revert "Remove png building"

This reverts commit fdd542de57316662f2c24ced0f6327d53e234650.

This is now fine to add the code back, so that it can be debugged when
times come.