Use Siunitx package for dealing with units
It is complicated :
[1] The International System of Units (si), https://www.bipm.org/en/
measurement-units/.
[2] International System of Units from nist, http://physics.nist.gov/cuu/Units/
index.html.

And one will never get it right (space or not, half-space?) nor
consistent so using that instead.

I am not sure how times and hours, when to write digits and when in
letter so I did not change much..

Did not touch the tables as ebooks on github actions seems to break when
you look at them funny.