Use sans-serif fonts for TikZ figures in website
Flowcharts were already sans-serif but the new one look weird with a
serif font on the website => let's make it sans by default.

For epub this a tough call but let's assume most people use sans there
as well, it's less weird anyway to have sans-serif pictures in a serif
text than the other way around.